requirements.txt
elasticsearch==5.1.0
PyYAML==3.11
aws-requests-auth==0.3.0